refactor: Http Response 관련 클래스 경로 이동
