+++ tomcat/src/main/java/org/apache/coyote/http11/response/HttpResponseHeader.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.response;
 
 import java.util.HashMap;
 import java.util.List;
@@ -10,7 +10,7 @@ public class HttpResponseHeader {
     public static final String TEXT_HTML_CHARSET_UTF_8 = "text/html;charset=utf-8";
     public static final String TEXT_CSS_CHARSET_UTF_8 = "text/css;charset=utf-8";
     private Map<String, String> headers;
- 
+
     public HttpResponseHeader(final String contentType, final String contentLength,
                               final String location, final String setCookie) {
         this.headers = new HashMap<>();
